Refactor PayrollTab helper classes
