+++ src/main/java/com/company/payroll/payroll/PayrollTab.java
@@ -37,10 +37,12 @@ import org.apache.pdfbox.pdmodel.common.PDRectangle;
 import org.apache.pdfbox.multipdf.PDFMergerUtility;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
-import javafx.concurrent.Worker;
 import javafx.scene.control.ButtonType;
 import javafx.scene.control.Dialog;
-import javafx.scene.control.ProgressBar;
+import com.company.payroll.payroll.CheckBoxListCell;
+import com.company.payroll.payroll.CheckListView;
+import com.company.payroll.payroll.PayrollHistoryEntry;
+import com.company.payroll.payroll.ProgressDialog;
 
 import java.io.*;
 import java.math.BigDecimal;
@@ -2220,235 +2222,4 @@ public class PayrollTab extends BorderPane implements
         alert.showAndWait();
     }
 
-    // Helper classes
-    private static class PayrollHistoryEntry {
-        final LocalDate date;
-        final String driverName;
-        final String truckUnit;
-        final int loadCount;
-        final double gross;
-        final double totalDeductions;
-        final double netPay;
-        final boolean locked;
-        
-        PayrollHistoryEntry(LocalDate date, String driverName, String truckUnit, int loadCount,
-                          double gross, double totalDeductions, double netPay, boolean locked) {
-            this.date = date;
-            this.driverName = driverName;
-            this.truckUnit = truckUnit;
-            this.loadCount = loadCount;
-            this.gross = gross;
-            this.totalDeductions = totalDeductions;
-            this.netPay = netPay;
-            this.locked = locked;
-        }
-    }
-
-    // CheckListView implementation
-	private static class CheckListView<T> extends ListView<T> {
-		private final CheckListViewModel checkModel;
-		
-		public CheckListView() {
-			this.checkModel = new CheckListViewModel();
-			setCellFactory(lv -> new CheckBoxListCell<>(
-				index -> checkModel.isSelected(index),
-				index -> checkModel.select(index),
-				index -> checkModel.clearSelection(index)
-			));
-		}
-		
-		public MultipleSelectionModel<T> getCheckModel() {
-			return checkModel;
-		}
-		
-		private class CheckListViewModel extends MultipleSelectionModel<T> {
-			private final ObservableList<Integer> selectedIndices = FXCollections.observableArrayList();
-			private final ObservableList<T> selectedItems = FXCollections.observableArrayList();
-			
-			@Override
-			public ObservableList<Integer> getSelectedIndices() { 
-				return FXCollections.unmodifiableObservableList(selectedIndices); 
-			}
-			
-			@Override
-			public ObservableList<T> getSelectedItems() { 
-				return FXCollections.unmodifiableObservableList(selectedItems); 
-			}
-			
-			@Override
-			public void selectIndices(int index, int... indices) {
-				select(index);
-				for (int i : indices) select(i);
-			}
-			
-			@Override
-			public void selectAll() {
-				for (int i = 0; i < getItems().size(); i++) select(i);
-			}
-			
-			@Override
-			public void selectFirst() { 
-				if (!getItems().isEmpty()) select(0); 
-			}
-			
-			@Override
-			public void selectLast() { 
-				if (!getItems().isEmpty()) select(getItems().size() - 1); 
-			}
-			
-			@Override
-			public void clearAndSelect(int index) { 
-				clearSelection(); 
-				select(index); 
-			}
-			
-			@Override
-			public void select(int index) {
-				if (index >= 0 && index < getItems().size() && !selectedIndices.contains(index)) {
-					selectedIndices.add(index);
-					selectedItems.add(CheckListView.this.getItems().get(index));
-				}
-			}
-			
-			@Override
-			public void select(T obj) {
-				int index = getItems().indexOf(obj);
-				if (index >= 0) select(index);
-			}
-			
-			@Override
-			public void clearSelection(int index) {
-				Integer idx = Integer.valueOf(index);
-				if (selectedIndices.contains(idx)) {
-					selectedIndices.remove(idx);
-					selectedItems.remove(getItems().get(index));
-				}
-			}
-			
-			@Override
-			public void clearSelection() {
-				selectedIndices.clear();
-				selectedItems.clear();
-			}
-			
-			@Override
-			public boolean isSelected(int index) {
-				return selectedIndices.contains(index);
-			}
-			
-			@Override
-			public boolean isEmpty() { 
-				return selectedIndices.isEmpty(); 
-			}
-			
-			@Override
-			public void selectPrevious() {
-				if (!selectedIndices.isEmpty()) {
-					int min = selectedIndices.stream().mapToInt(Integer::intValue).min().orElse(0);
-					if (min > 0) select(min - 1);
-				}
-			}
-			
-			@Override
-			public void selectNext() {
-				if (!selectedIndices.isEmpty()) {
-					int max = selectedIndices.stream().mapToInt(Integer::intValue).max().orElse(0);
-					if (max < getItems().size() - 1) select(max + 1);
-				}
-			}
-		}
-	}
-
-    private static class CheckBoxListCell<T> extends ListCell<T> {
-        private final CheckBox checkBox = new CheckBox();
-        private final java.util.function.Function<Integer, Boolean> isSelected;
-        private final java.util.function.Consumer<Integer> onSelect;
-        private final java.util.function.Consumer<Integer> onDeselect;
-        
-        public CheckBoxListCell(java.util.function.Function<Integer, Boolean> isSelected,
-                               java.util.function.Consumer<Integer> onSelect,
-                               java.util.function.Consumer<Integer> onDeselect) {
-            this.isSelected = isSelected;
-            this.onSelect = onSelect;
-            this.onDeselect = onDeselect;
-            
-            setContentDisplay(ContentDisplay.LEFT);
-        }
-        
-        @Override
-        protected void updateItem(T item, boolean empty) {
-            super.updateItem(item, empty);
-            if (empty || item == null) {
-                setGraphic(null);
-                setText(null);
-            } else {
-                checkBox.setSelected(isSelected.apply(getIndex()));
-                checkBox.setOnAction(e -> {
-                    if (checkBox.isSelected()) {
-                        onSelect.accept(getIndex());
-                    } else {
-                        onDeselect.accept(getIndex());
-                    }
-                });
-                
-                // Format the text based on the item type
-                if (item instanceof Load) {
-                    Load load = (Load) item;
-                    setText(String.format("Load #%s - %s to %s ($%.2f)", 
-						load.getLoadNumber(),
-						load.getPickupCity() + ", " + load.getPickupState(),
-						load.getDeliveryCity() + ", " + load.getDeliveryState(),
-						load.getDriverRate()));
-                } else {
-                    setText(item.toString());
-                }
-                
-                setGraphic(checkBox);
-            }
-        }
-    }
-	    // Inner class for progress dialogs
-    private static class ProgressDialog<T> extends Dialog<T> {
-        private final ProgressBar progressBar;
-        private final Label messageLabel;
-        
-        public ProgressDialog(Task<T> task) {
-            setTitle("Processing");
-            setHeaderText("Please wait...");
-            
-            VBox content = new VBox(10);
-            content.setPadding(new Insets(20));
-            
-            progressBar = new ProgressBar();
-            progressBar.setPrefWidth(300);
-            progressBar.progressProperty().bind(task.progressProperty());
-            
-            messageLabel = new Label();
-            messageLabel.textProperty().bind(task.messageProperty());
-            
-            content.getChildren().addAll(messageLabel, progressBar);
-            getDialogPane().setContent(content);
-            
-            // Only allow closing when task is done
-            task.stateProperty().addListener((obs, oldState, newState) -> {
-                if (newState == Worker.State.SUCCEEDED || 
-                    newState == Worker.State.FAILED || 
-                    newState == Worker.State.CANCELLED) {
-                    close();
-                }
-            });
-            
-            // Disable close button while running
-            getDialogPane().getButtonTypes().add(ButtonType.CANCEL);
-            Node cancelButton = getDialogPane().lookupButton(ButtonType.CANCEL);
-            cancelButton.disableProperty().bind(task.runningProperty().not());
-            
-            setResultConverter(dialogButton -> {
-                if (dialogButton == ButtonType.CANCEL) {
-                    task.cancel();
-                }
-                return null;
-            });
-        }
-    }
 }
